+++ src/libaktualizr/primary/aktualizr.cc
@@ -68,6 +68,12 @@ void Aktualizr::UptaneCycle() {
 
   uptane_client_->uptaneInstall(download_result.updates);
   uptane_client_->putManifest();
+  if (uptane_client_->isInstallCompletionRequired()) {
+    // If there are some pending updates then effectively either reboot (ostree) or aktualizr restart (fake pack mngr)
+    // is required to apply the update(s)
+    LOG_INFO << "About to exit aktualizr so the pending updates can be applied after reboot";
+    Shutdown();
+  }
 }
 
 std::future<void> Aktualizr::RunForever() {
@@ -77,6 +83,7 @@ std::future<void> Aktualizr::RunForever() {
       UptaneCycle();
       std::this_thread::sleep_for(std::chrono::seconds(config_.uptane.polling_sec));
     }
+    uptane_client_->completeInstall();
   });
   return future;
 }