+++ src/libaktualizr/primary/sotauptaneclient.cc
@@ -1193,6 +1193,17 @@ void SotaUptaneClient::campaignAccept(const std::string &campaign_id) {
   report_queue->enqueue(std_::make_unique<CampaignAcceptedReport>(campaign_id));
 }
 
+bool SotaUptaneClient::isInstallCompletionRequired() {
+  bool force_install_completion = (hasPendingUpdates() && config.uptane.force_install_completion);
+  return force_install_completion;
+}
+
+void SotaUptaneClient::completeInstall() {
+  if (isInstallCompletionRequired()) {
+    package_manager_->completeInstall();
+  }
+}
+
 bool SotaUptaneClient::putManifestSimple() {
   // does not send event, so it can be used as a subset of other steps
   auto manifest = AssembleManifest();