+++ src/libaktualizr/primary/sotauptaneclient.h
@@ -106,6 +106,9 @@ class SotaUptaneClient {
   void installationComplete(const std::shared_ptr<event::BaseEvent> &event);
   result::CampaignCheck campaignCheck();
   void campaignAccept(const std::string &campaign_id);
+  bool hasPendingUpdates();
+  bool isInstallCompletionRequired();
+  void completeInstall();
   LazyTargetsList allTargets();
 
   static Uptane::Targets getTrustedDelegation(const Uptane::Role &delegate_role, const Uptane::Targets &parent_targets,
@@ -118,6 +121,7 @@ class SotaUptaneClient {
   FRIEND_TEST(Aktualizr, FullMultipleSecondaries);
   FRIEND_TEST(Aktualizr, CheckNoUpdates);
   FRIEND_TEST(Aktualizr, DownloadWithUpdates);
+  FRIEND_TEST(Aktualizr, AutoRebootAfterUpdate);
   FRIEND_TEST(Uptane, AssembleManifestGood);
   FRIEND_TEST(Uptane, AssembleManifestBad);
   FRIEND_TEST(Uptane, InstallFake);
@@ -154,7 +158,6 @@ class SotaUptaneClient {
   void sendMetadataToEcus(const std::vector<Uptane::Target> &targets);
   std::future<bool> sendFirmwareAsync(Uptane::SecondaryInterface &secondary, const std::shared_ptr<std::string> &data);
   std::vector<result::Install::EcuReport> sendImagesToEcus(const std::vector<Uptane::Target> &targets);
-  bool hasPendingUpdates();
   void sendDownloadReport();
 
   bool putManifestSimple();